backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy==1.23.5
torch==2.0.1
torchvision==0.15.2
werkzeug==3.0.1
PyJWT==2.8.0
mysql-connector-python==8.0.29
twilio==8.10.0
